Implemented: Added a separate component containing a search bar and sorting options for landing pages in the admin section(#631)
