+++ src/views/Closed.vue
@@ -15,15 +15,7 @@
 
     <ion-content ref="contentRef" :scroll-events="true" @ionScroll="enableScrolling()" id="filter">
       <div class="header searchbar">
-        <ion-searchbar v-model="query.queryString" @keyup.enter="updateQueryString('queryString', $event.target.value)" />
-        <ion-item lines="none">
-          <ion-icon slot="start" :icon="swapVerticalOutline" />
-          <ion-select :label="translate('Sort by')" :value="query.sortBy" @ionChange="updateQuery('sortBy', $event.detail.value)" interface="popover">
-            <ion-select-option value="createdDate desc">{{ translate("Created date") }}</ion-select-option>
-            <ion-select-option value="dueDate desc">{{ translate("Due date") }}</ion-select-option>
-            <ion-select-option value="countImportName asc">{{ translate("Alphabetic") }}</ion-select-option>
-          </ion-select> 
-        </ion-item>
+        <SearchBarAndSortBy @emitUpdateQueryString="updateQueryString" />
         <ion-item lines="full">
           <ion-icon slot="start" :icon="listOutline"/>
           <ion-label>{{ translate("Counts closed") }}</ion-label>
@@ -120,6 +112,7 @@ import Filters from "@/components/Filters.vue"
 import store from "@/store";
 import { getCycleCountStats, getDateWithOrdinalSuffix, getFacilityName } from "@/utils";
 import DownloadClosedCountModal from "@/components/DownloadClosedCountModal.vue";
+import SearchBarAndSortBy from "@/components/SearchBarAndSortBy.vue";
 import router from "@/router"
 
 const cycleCounts = computed(() => store.getters["count/getCounts"])
@@ -152,8 +145,7 @@ function enableScrolling() {
   }
 }
 
-async function updateQueryString(key: string, value: any) {
-  await store.dispatch("count/updateQueryString", { key, value })
+async function updateQueryString() {
   await Promise.allSettled([fetchClosedCycleCounts(), store.dispatch("count/fetchClosedCycleCountsTotal")])
 }
 
@@ -230,6 +222,6 @@ ion-content {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   align-items: center;
-  gap: 10px;
+  gap: var(--spacer-xs);
 }
 </style>
