+++ src/views/PendingReview.vue
@@ -14,15 +14,7 @@
 
     <ion-content ref="contentRef" :scroll-events="true" @ionScroll="enableScrolling()" id="filter">
       <div class="header searchbar">
-        <ion-searchbar v-model="query.queryString" @keyup.enter="updateQueryString('queryString', $event.target.value)" />
-        <ion-item lines="none">
-          <ion-icon slot="start" :icon="swapVerticalOutline" />
-          <ion-select :label="translate('Sort by')" :value="query.sortBy" @ionChange="updateQuery('sortBy', $event.detail.value)" interface="popover">
-            <ion-select-option value="createdDate desc">{{ translate("Created date") }}</ion-select-option>
-            <ion-select-option value="dueDate desc">{{ translate("Due date") }}</ion-select-option>
-            <ion-select-option value="countImportName asc">{{ translate("Alphabetic") }}</ion-select-option>
-          </ion-select> 
-        </ion-item>
+        <SearchBarAndSortBy @emitUpdateQueryString="updateQueryString" />
       </div>
       <p v-if="!cycleCounts.length" class="empty-state">
         {{ translate("No cycle counts found") }}
@@ -75,6 +67,7 @@ import store from "@/store"
 import router from "@/router"
 import Filters from "@/components/Filters.vue"
 import { getCycleCountStats, getDateWithOrdinalSuffix, getDerivedStatusForCount, getFacilityName } from "@/utils"
+import SearchBarAndSortBy from "@/components/SearchBarAndSortBy.vue";
 
 const cycleCounts = computed(() => store.getters["count/getCounts"])
 const isScrollable = computed(() => store.getters["count/isCycleCountListScrollable"])
@@ -104,8 +97,7 @@ function enableScrolling() {
   }
 }
 
-async function updateQueryString(key: string, value: any) {
-  await store.dispatch("count/updateQueryString", { key, value })
+async function updateQueryString() {
   fetchPendingCycleCounts();
 }
